minim.opm: Patch `\newcatcodetable` before loading minim-alloc (#154)
In the latest release minim actually uses `\newcatcodetable` in
minim-alloc.tex. This means that it still uses the standard OpTeX definition
(which doesn't initialize the catcode table), but it expects the
initialization.

We solve this by doing the patch to `\newcatcodetable` _before_ loading
minim-alloc.tex. It should have probably been done this way right from
the beginning, it just didn't manifest.