Improve error handling for registering callbacks
When we initially introduced support for luatexbase / ltluatex style
callbacks (luatexbase.add_to_callback, etc.), we introduced also some
checking and reported some errors with `err`, though we were not as
thorough as the original LaTeX code, because we mainly aimed to provide
compatibility with existing packages, which already were using the
functions correctly, as they had to go through LaTeX's more extensive
checking.

As time progressed, we in OpTeX added more uses of the callbacks, and
OpTeX users themselves have been using the mechanism, since in OpTeX we
don't allow direct registering of callbacks, it is always necessary to
go through `callback.add_to_callback`. However, this mean that some uses
could not obey the "rules", while we were expecting some invariants to
hold true.

One such problem manifested for callbacks added without descriptions.
Internally we expected that each callback function would be registered
with a description, and that a description wouldn't be used more than
once for a single callback name.

This was not an immediate problem when _adding_ the callbacks, but could
lead to problems when removing from callbacks, since that is where we
assume some consistency in the registered functions, and descriptions.
If the description was `nil` (i.e. not passed) the internal list of
functions and descriptions could drift away, causing callbacks to be
deregistered prematurely, as the condition for deregistration is that
there are no descriptions left.

The solution implemented here is a more thorough checking of arguments
and consistency.

As part of the consistency checking, I found that our use of
"finish_pdffile" callback itself didn't conform to the invariants (had
no description), further suggesting, that such mistakes are easy to make
accidentally. That is also fixed here.