Allow multiple callbacks with the same description
This cautious check for multiple entries was added in last commit, and
it turned out to be problematic with our `\mathsbon` callback, which can
be activated multiple times in nested `\load`s (see #180).

While it seems better to me, to be proactive about the checks, LuaLaTeX
also seems to allow duplicate entries, and results in same behavior when
removal is involved (the first registered is also the first to be
removed). So, let's just remove the check.