use post_mlist_to_hlist_filter and add the function only once in \enablefic
Since 3b46d2a a description cannot be used if it is already in the callback,
so currently using \enablefic a second time will throw an error.
Redefining \enablefic after its first call to only set \finalitalcorr
to 1 will solve it.

The mlist_to_hlist callback is exclusive so only one function can
be added to it. Since \enablefic only use the horizontal list and
not the horizontal list, it can be added to post_mlist_to_hlist_filter
which can have few functions added.

In addition this commit optimize the function a bit, and avoid using
a function name in the global scope.