Conservatively always set stroke color for glyphs
In an attempt to fix outline font coloring (#204), we started checking a
font "mode" of each glyph to see if it is an outline mode (which
requires stroke color). Unfortunately, this turned out to regress
performance too much.

While there are other solutions that we can explore in the future, for
now we will just always set stroke color for glyphs.

This removes some of the size improvements [1] that we got with colors
based on attributes, but it shouldn't be as noticeable as the
performance regression.

[1]: https://github.com/olsak/OpTeX/pull/57#issuecomment-869160661